simplify: Fix area computation for quadricFromAttributes
In most places in meshoptimizer, area is a short-hand for "2x area" and
is used exclusively as a normalization factor so the extra 2x cancels
out. In attribute metric, it is used in a way that is user-visible and
affects the results as it changes the attribute weighting behavior.

We could probably compute the real area everywhere but for simplicity
let's just update this here as it is the only place where this
distinction matters.