demo: Implement simplification debug overlay
To get data for debug overlay, we run simplification process again and
output loop/kind metadata. To minimize the changes to the rendering
code, we parse the index buffer and generate extra triangles (for now
using the same color) for any loop edge, and use three's material groups
to render that with a separate material. Conveniently, even if the
original mesh has vertex colors, these get ignored by MeshBasicMaterial
so this works properly, and is compatible with skinning because we use
wireframe on degenerate triangles (three doesn't support skinned lines
otherwise).

This is a little crude and will be refined in the future but gets the
job done.