demo: Add basic (optional) welding to the simplifier
Without this, some glTF models can't be properly simplified because of
export or conversion issues resulting in duplicate vertices with equal
or slightly-different normals or tangents. Ideally in the future
simplifier (with attributes?) should be more tolerant or we should add a
pre-process step, but for now we can use mergeVertices from three.js.

The tolerance is set to be reasonable for normals; unfortunately
mergeVertices does not allow to specify a separate value for positions
or UVs (which we would rather keep unwelded).