demo: Add a test for simplification along attribute seam
This test is intended mostly to track future improvements to the error
behavior along the seam boundary. Right now, attribute error is not
correctly computed along the seam boundary which prevents some collapses
and can result in artificially inflated resulting error. We also seem to
have some issues with positional errors for collapses along the seam
which the test also demonstrates.