simplify: Minor refactoring of performEdgeCollapses
To avoid mistakes, extract kind into a variable (we're only concerned
with source vertex kind for classification since that's what determines
the wedges we need to remap), and also add a few comments.

Notably, we probably don't need to worry about attribute quadrics for
complex vertices long term; the intention for complex classification has
been to treat vertices with attributes that are similar enough as one,
so this probably can function similarly to ranking where we're okay with
ignoring the topology somewhat.

Also, while we're here, adjust complex collapses to collapse to i1, not
r1. This code is not used right now but collapsing to r1 feels slightly
wrong since in some cases we'd be collapsing to a vertex that was never
connected to any wedge in the complex.