demo: Adjust simplifyAttr test wrt numerical stability
There is a little bit too much ambiguity in the test right now as there
are multiple ways to tessellate the final output that are all
equivalent; we also forgot to adjust the weights after the metric
changed, so the attribute impact is insignificant.

Hopefully a higher attribute weight and more variation in height will
make this behave the same across platforms.