demo: Add tunable textureWeight for simplify.html
This is fairly rarely useful as UV distortion is mostly taken care of
via seam classification but it can be helpful for testing. For now we
limit weight to 100; on a 1024x1024 texture a weight of 100 corresponds
to the error of ~10 pixels being as bad as just removing geometry. This
is likely abnormally high in practice but the default here is 0 anyway.