gltfpack: For now always use error=1e-1f for simplifySloppy
Before this we used the error specified via command line for
simplifySloppy when -sa was specified. It probably makes more sense to
keep the hardcoded high error; the documentation says that this option
disregards quality, and this way the behavior is cleaner wrt defaults
since -se 0.01 is equivalent to not specifying it. If we need control
over aggressive error limit we can revert this commit in the future.