indexgenerator: Introduce initial version of provoking vertex algorithm
This algorithm restructures the index buffer to satisfy two constraints:

- indices[3 * tri + 0] == tri
- reorder[indices[x]] refers to the original vertices

This index buffer requires a vertex shader that performs the indirection
through reorder[], and because this construction requires at least one
vertex per triangle it can be less efficient to render, however this
allows passing the vertex id to the fragment shader using nointerpolate
attribute (via provoking vertex), which is equal to primitive id.

This is important for performance of visibility buffer based renderers
on hardware where primitive id can't be accessed efficiently.

This is an initial implementation loosely based on SIGGRAPH 2024
"Variable Rate Shading with Visibility Buffer Rendering" talk; we're
missing some features like valence based prioritization, but because the
algorithm works on the entire mesh it already produces better results
than the cluster-based version suggested in the talk.