indexgenerator: Use vertex valence to prioritize provoking vertex
Instead of just using any free vertex in a triangle to be the provoking
vertex, we now look at vertex valence and choose the free vertex with
minimum valence. This reduces the probability that a future triangle
will have all three vertices become used, which reduces the number of
extra vertices we end up creating (by about ~half on smooth meshes).