indexgenerator: Clean up provoking remapping pass
When all three vertices are used we don't really care which one we take
to be the provoking one, as neither can be reused by future triangles
anyway.

This makes it possible to streamline the logic following the triangle
rotation, as we *must* map the provoking vertex to the new index
regardless of the case. Hopefully the comments make the flow reasonably
clear to the future reader.