indexgenerator: Use 8-bit valence with wrap-around for performance
Valence initialization and use is a significant cost relative to the
rest of the work that happens during provoking index buffer generation.
We use a 32-bit buffer because that guarantees correct valence, but in
reality most meshes have valence that's less than 10, and even amid
outliers it's very rare to see a value that is 256 or more.

Moreover, since valence is only used for ordering heuristics, it is
completely safe to compute incorrect valences for outliers, and this
will only marginally affect the output. Using an 8-bit buffer here makes
provoking index buffer generation ~10% faster.

The same change with the same reasoning can be made to the stripifier;
while it's less valuable from the performance point of view, we make it
for consistency as the same logic applies.