Update simplification documentation to remove shadowIndexBuffer
While both the header comments and README recommented
meshopt_generateShadowIndexBuffer for pre-welding during simplification,
this guidance was incorrect: simplifying a shadow index buffer is
currently subject to some of the same restrictions as the main index
buffer because the classification is partially based on the contents of
the entire vertex buffer including unreferenced vertices.

This could change in the future, and can be worked around via
meshopt_SimplifySparse, but because it's not clear if this is even a
preferred route going forward, for now remove all incorrect suggestions
and keep meshopt_generateShadowIndexBuffer for its narrow use of
generating optimal index buffers for depth only rendering.