Update README.md
Add documentation for all three algorithms that can be used to generate
special versions of an index buffer that are useful for very specific
rendering algorithms.

Note that meshopt_generateProvokingIndexBuffer is experimental; while we
typically avoid documenting APIs until they are stabilized, the
interface here is unlikely to change even if the implementation does.