gltfpack: Use quantized normal/tangents for reindexing by default
While reindexing full precision streams is usually sufficient to get a
good non-redundant index buffer, some meshes go through export or
processing pipelines that introduce imprecision in individual normals
or tangents, creating artificial seams. This increases the storage size
and decreases rendering efficiency, in addition to creating issues for
simplification.

This can be solved with a tolerance-based weld pass, but for now we can
simply use quantized normals/tangents during reindexing. Note that this
by itself does not reduce precision of individual vertices, so in theory
we can be *more* aggressive here with quantization bits - for now we
always use 8 bits which should not cause issues even if the requested
normal bits is 16.