simplify: Adjust meshopt_simplifyPoints color errors to be more uniform
Since the error for points is pos + color_weight * col, it is hard to
tweak color weight to be scale-independent. We had a similar issue in
the edge collapse metric, but here the issue is complicated by the fact
that it's incorrect to compute errors as pos * (1 + col), since multiple
points can have the same position but the color delta must be used for
picking the right replacement.

As a compromise, we scale the color weight by the grid scale; since the
maximum position error along any axis is the grid scale, this makes the
errors compatible, and makes it easier to tweak the weight: similarly to
edge collapse, a weight of 1 is a good default; at a weight of 1 a
distance of 1 in color is equivalent to a worst case positional delta,
so a point like this will only be selected if there is no closer option.