demo: Rework cluster processing to work with stable cluster indices
This is required to be able to properly compute and update LOD
information; additionally, this reduces the amount of data copying we
have to do for indices and will make it easier to reintroduce meshlet
local indexing in the future.

As part of this, simplify() also returns indices instead of a cluster,
as it would not make sense to compute meshlet local indices here.