demo: Implement cluster LOD bounds tracking
We now compute LOD bounds for each cluster as well as the parent
information that propagates through DAG and should make it possible to
select a DAG cut just based on bounds alone.

This also makes it possible to compute the lowest LOD (as well as any
other LOD!) from cluster data alone without relying on stuck_triangles
diagnostics which has been reworked to just analyze the current LOD
level.