demo: Add DAG error validation to check monotonicity
For correct hierarchical LOD selection we need to maintain the bounds
monotonicity: any parent cluster needs to have error >= any child
cluster from any viewpoint. This is something that we don't currently
get right because our merged sphere might not cover all child spheres;
for now we will print an error if this happens, but this code might be
removed in the future.