demo: Reinsert stuck clusters into the queue
This means that small clusters that end up being too small to merge or
edge-locked enough so that simplification is not effective may get
another chance further into the process to get merged with other
clusters. This is generally beneficial for quality although sometimes
results in worse results and definitely results in slower processing as
the stuck triangles keep being reevaluated on every pass.