demo: Rework DAG thresholding to be more lenient
Especially Mhen Metis triangle clustering is used, we often get much
smaller clusters in the initial split, because it gets to 129 triangles
and splits that into 64+65. Then partioning may either take three
clusters with sizes a little under 128, or four clusters two of which
are 64/65, and the resulting cluster will be too small to merge.

Instead we remove the merge criteria outright if we have two clusters to
merge, and replace it with percentage reduction. Right now the
percentage is very lenient; it would not be ideal to have every single
level to just be a 85% reduction. However, this can be controlled on a
macro level and ideally we want to prevent the processing getting stuck
if possible.