demo: Implement connectivity analysis for meshlets
Since whether meshlet is connected or not is a very important criteria
for how well it will do in a hierarchical clusterization process, we now
compute the number of connected components in meshlet demo and count the
number of meshlets with more than one.

Notably, for now we do this analysis using indices; this means we do
connectivity analysis on original, non-positional, topology, and on some
geometry that looks visually connected we will naturally have many
disconnected meshlets.