demo: Add auto-LOD testing mode to simplify demo
Instead of using a controlled ratio with a fixed threshold, we can now
use the distance to the object's surface to scale the expected error and
simplify as much as we can with that limit.

For now we don't carefully account for multi-mesh scenes where the
difference in submesh scale and size would require adjusting the allowed
error, and the scene radius is assumed to be 1 because that's what we
resize it to. This is sufficient for testing this but may improve later.