tools: Add simplifyfuzz to stress test simplifier
The code for classifying topology and using the resulting data is
getting a little more complicated and critical than it used to be; while
we can test it on a wide set of meshes with various strange topology, it
would also help to be able to fuzz it.

We build a vertex buffer that will occasionally have redundancy in it
due to random collisions, which will generate seam/locked vertices, and
use the rest of the fuzz input as an index buffer.